- abstracting booking form fields into separate components - creating a separate form for new bookins and edit bookings
