+++ app/routes/_layout+/bookings.$bookingId.tsx
@@ -13,13 +13,22 @@ import { dynamicTitleAtom } from "~/atoms/dynamic-title-atom";
 import { BookingStatusBadge } from "~/components/booking/booking-status-badge";
 import { CheckinIntentEnum } from "~/components/booking/checkin-dialog";
 import { CheckoutIntentEnum } from "~/components/booking/checkout-dialog";
-import { BookingFormSchema } from "~/components/booking/form";
+import { BookingFormSchema } from "~/components/booking/forms/forms-schema";
 import { BookingPageContent } from "~/components/booking/page-content";
+import { TimeRemaining } from "~/components/booking/time-remaining";
 import ContextualModal from "~/components/layout/contextual-modal";
 import ContextualSidebar from "~/components/layout/contextual-sidebar";
 import Header from "~/components/layout/header";
 import type { HeaderData } from "~/components/layout/header/types";
+import { Button } from "~/components/shared/button";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from "~/components/shared/tooltip";
 import { db } from "~/database/db.server";
+import { useDisabled } from "~/hooks/use-disabled";
 import { hasGetAllValue } from "~/hooks/use-model-filters";
 import {
   archiveBooking,
@@ -765,8 +774,16 @@ export default function BookingPage() {
         subHeading={
           <div key={booking.status} className="flex items-center gap-2">
             <BookingStatusBadge status={booking.status} />
+            <TimeRemaining
+              from={booking.from!}
+              to={booking.to!}
+              status={booking.status}
+            />
           </div>
         }
+        slots={{
+          "right-of-title": <AddToCalendar />,
+        }}
       />
 
       <div>
@@ -777,3 +794,36 @@ export default function BookingPage() {
     </div>
   );
 }
+
+const AddToCalendar = () => {
+  const disabled = useDisabled();
+  const { booking } = useLoaderData<typeof loader>();
+  const isArchived = booking.status === BookingStatus.ARCHIVED;
+  return (
+    <div className="ml-auto">
+      <TooltipProvider delayDuration={100}>
+        <Tooltip>
+          <TooltipTrigger asChild>
+            <Button
+              to={`cal.ics`}
+              download={true}
+              reloadDocument={true}
+              disabled={disabled || isArchived}
+              variant="secondary"
+              icon="calendar"
+            >
+              Add to calendar
+            </Button>
+          </TooltipTrigger>
+          <TooltipContent side="bottom">
+            <p className="text-xs">
+              {disabled
+                ? "Not possible to add to calendar due to booking status"
+                : "Download this booking as a calendar event"}
+            </p>
+          </TooltipContent>
+        </Tooltip>
+      </TooltipProvider>
+    </div>
+  );
+};
